server/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-dotenv
Pillow
sqlalchemy
easyocr
opencv-python-headless
